+++ src/wallet/inscribe/batch_entry.rs
