+++ src/wallet/inscribe/batch_file.rs
